Several changes made in the graphics layer so that the camera scrollbox may be displayed on screen.
ScrollBox is a new class in the /Model directory. This class will index fixed Skia coordinates to modifiable game logic coordinates, and will be how the camera follows a set focus and enacts scrolling.

The GPVM has a ScrollBox object.

The DeviceScreen has a new SKRect property corresponding to the SK-side of gpvm's scrollbox.

Screen class has further encapsulated its internal members in a ScreenInfo class, which is easier to pass / trivially copyable compared to the GameScreen which contains sprites, tiles, render artifacts, etc..

Passing objects of ScreenInfo is how information quickly gets from OnSizeAllocated to update the ScrollBox's skia coordinates, and back to the device screen to update its scrollbox coordinates.

ScrollBox has to wait until a level has been loaded to be initialized. For this reason its Update() function is called in GPVM's constructor following level load, and OnSizeAllocated events.

GameScreen also has a Troubleshooting property to display different artfiacts to assist in debugging.